Fixed a bunch of simple signed/unsigned warnings in libgens and libzomg.
log_msg.c: Set the last byte of out_msg to NULL after calling vsnprintf().
It's not needed for the initial snprintf(), since we're checking the
return value of that function call anyway.

+++ src/libgens/MD/EmuMD.cpp
@@ -64,7 +64,7 @@ EmuMD::EmuMD(Rom *rom)
 	}
 	
 	// Check the ROM size.
-	if ((rom->romSize() == 0) || (rom->romSize() > sizeof(M68K_Mem::Rom_Data)))
+	if ((rom->romSize() == 0) || (rom->romSize() > (int)sizeof(M68K_Mem::Rom_Data)))
 	{
 		// ROM is either empty or too big.
 		// TODO: Set an error code.
